services/server/requirements.txt
fastapi[standard]
psycopg2-binary
pydantic-settings
sqlalchemy
passlib